feat: Add service and job for refreshing materialized views using swap (#70)
Introduces a new `swap` refresh strategy for materialized views to enable near-zero downtime updates.

This strategy works by creating a new temporary view in the background. Once the new view is populated, it is atomically swapped with the original view inside a transaction. This process minimizes the time the view is unavailable for queries.

The service also handles recreating any declared unique indexes on the new view after the swap.