feat: Add rake tasks to delete materialized views (#79)
* feat: Add rake tasks to delete materialized views

Introduces new Rake tasks for enqueuing jobs to delete materialized views. This provides a way to manage the lifecycle of views directly from the command line.

The new tasks are:
- `mat_views:delete_by_name`
- `mat_views:delete_by_id`
- `mat_views:delete_all`

All tasks support a `cascade` option to control the `DROP MATERIALIZED VIEW` behavior and include a confirmation prompt to prevent accidental deletion, which can be skipped.

* Clarify error messages for rake tasks

Improves the error messages for `_by_id` rake tasks when the `definition_id` parameter is missing.

The new messages include the full task name, making it clearer to the user which task is failing and why.