lang: Add localization: en-NP (#196)
Dynamically discovers available locales by reading the contents of the `config/locales` directory. This removes the need for hardcoded lists of locales in the engine, CI workflows, and tests.

The CI workflow for e2e tests is updated to use a matrix strategy based on the discovered locales. This automatically generates screenshot artifacts and pull request comments for any new locales without requiring manual workflow updates.

This change also introduces a new localization for English (Nepal) (en-NP) and adds `en` as a base locale.