feat: Introduce server-side datatables (#323)
Using datatables in both definitions and runs tables now gives a robust way to show the information.
With search, sort, filter and pagination the UI is more inviting.

Replaces the static definitions table with a server-side powered datatable, enabling multi-column sorting and full-text search. This improves the user experience by allowing for easier navigation and filtering of materialized view definitions and runs.

This change also includes:
- A major refactoring of the CSS to use a design token system for consistency.
- New Stimulus controllers to manage datatable state and theme application.
- Scope additions to the model for sorting and searching capabilities.